+++ bakedanuki/bakedanuki-util/python/bd_util/ui/__init__.py
@@ -54,6 +54,7 @@ from .binding import (
     SetBoolCommand,
 )
 from .settings_path import SettingsPath
+from .radio_button_sweep import RadioButtonSweep
 from .ui_state import UiStateManager
 from .window import WindowController
 from .window_state import (
@@ -63,6 +64,7 @@ from .window_state import (
 )
 
 __all__ = [
+    "RadioButtonSweep",
     "EnumBinding",
     "EnumComboBox",
     "EnumDefinition",